feat!: remove unnecessary dependencies from RPM spec
On most RPM-based systems, mkdir and touch are provided by the coreutils package,
which is a fundamental part of the OS. It's highly unlikely for a system to be missing these.

On the other hand this currently breaks on amazonlinux because their coreutils package
does not provide these commands explicitly.