Create archive tar with appropriate flags rather than chown
In order to prevent file ownership issues when extracting the tarball,
we were chowning all files to root before creating the tarball. Using
chown on Windows in Cygywin in GitHub Actions is dicey, so instead, we
use the appropriate tar flags to ensure all files have ownership by 0:0.
When extracted, if extracting as a non-root user, since a non-root user
can't create files they don't have access to, that user will end up
owning the files.